[ui-layer-refactor] refactor : Subway관련 adapter들 common 패키지로 이동
