+++ app/src/main/java/com/hegunhee/subwayarrivalinfoapp/ui/common/subway/SubwayInfoAdapter.kt
@@ -1,4 +1,4 @@
-package com.hegunhee.subwayarrivalinfoapp.ui.main
+package com.hegunhee.subwayarrivalinfoapp.ui.common.subway
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
@@ -7,10 +7,11 @@ import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.hegunhee.subwayarrivalinfoapp.data.entity.SubwayInfoEntity
 import com.hegunhee.subwayarrivalinfoapp.databinding.ItemSubwayInfoBinding
+import com.hegunhee.subwayarrivalinfoapp.ui.main.MainFragmentActionHandler
 
 class SubwayInfoAdapter(
     private val eventHandler: MainFragmentActionHandler
-) : ListAdapter<SubwayInfoEntity,SubwayInfoAdapter.MainViewHolder>(diffUtil) {
+) : ListAdapter<SubwayInfoEntity, SubwayInfoAdapter.MainViewHolder>(diffUtil) {
 
     inner class MainViewHolder(private val binding : ItemSubwayInfoBinding) : RecyclerView.ViewHolder(binding.root){
 
@@ -29,12 +30,13 @@ class SubwayInfoAdapter(
     override fun onBindViewHolder(holder: MainViewHolder, position: Int) {
         holder.bind(getItem(position))
     }
-}
+    companion object {
+        private val diffUtil = object : DiffUtil.ItemCallback<SubwayInfoEntity>(){
+            override fun areItemsTheSame(oldItem: SubwayInfoEntity, newItem: SubwayInfoEntity): Boolean =
+                oldItem.subwayName == newItem.subwayName
 
-internal object diffUtil : DiffUtil.ItemCallback<SubwayInfoEntity>(){
-    override fun areItemsTheSame(oldItem: SubwayInfoEntity, newItem: SubwayInfoEntity): Boolean =
-        oldItem.subwayName == newItem.subwayName
-
-    override fun areContentsTheSame(oldItem: SubwayInfoEntity, newItem: SubwayInfoEntity): Boolean =
-        oldItem == newItem
+            override fun areContentsTheSame(oldItem: SubwayInfoEntity, newItem: SubwayInfoEntity): Boolean =
+                oldItem == newItem
+        }
+    }
 }
